Bug 1809918 - Make WidgetPointerHelper use int32_t for tiltX/tiltY; r=smaug
For Windows, the `tiltX`/`tiltY` are from `POINTER_PEN_INFO::tiltX`/`tiltY`,
https://learn.microsoft.com/en-us/windows/win32/api/winuser/ns-winuser-pointer_pen_info,
which are `INT32`.

For Mac, the `tiltX`/`tiltY` are from `NSEvent::tilt`, https://developer.apple.com/documentation/appkit/nsevent,
which is a `CGPoint`, https://developer.apple.com/documentation/corefoundation/cgpoint,
the properties are `Double`, using `int32_t` still makes more sense.

Differential Revision: https://phabricator.services.mozilla.com/D166725

+++ widget/cocoa/nsChildView.mm
@@ -3338,8 +3338,8 @@ static gfx::IntPoint GetIntegerDeltaForEvent(NSEvent* aEvent) {
     MOZ_ASSERT(aOutGeckoEvent->mPressure >= 0.0 && aOutGeckoEvent->mPressure <= 1.0);
   }
   aOutGeckoEvent->mInputSource = dom::MouseEvent_Binding::MOZ_SOURCE_PEN;
-  aOutGeckoEvent->tiltX = lround([aPointerEvent tilt].x * 90);
-  aOutGeckoEvent->tiltY = lround([aPointerEvent tilt].y * 90);
+  aOutGeckoEvent->tiltX = (int32_t)lround([aPointerEvent tilt].x * 90);
+  aOutGeckoEvent->tiltY = (int32_t)lround([aPointerEvent tilt].y * 90);
   aOutGeckoEvent->tangentialPressure = [aPointerEvent tangentialPressure];
   // Make sure the twist value is in the range of 0-359.
   int32_t twist = fmod([aPointerEvent rotation], 360);
